`Mtgapi::Card` object
Right now it just inherits from `Hashi::Mash`. I wouldn't have bothered
to wrap a `Mash` in my own object but wanted to do it now so there's a
place to put convenience methods for accessing data from the api.

I.E. `Card#flavor` returns an array with a quote as the first element
and an attribution as the seccond. The Card object will probably end
up with methods to access those without calling `first` or `last` on `flavor`.